Move FieldSchema et al. to Schema/Fields
Move FromSqlValues, HasDefaultFieldSchema and FieldSchema
to Schema/Fields in order to avoid orphan instances.
